+++ lib/src/widgets/components/calculate_second_pattern_page.dart
@@ -5,17 +5,17 @@ import 'package:sugars_check/src/blocs/sugars_bloc.dart';
 class CalculateSecondPatternPage extends StatelessWidget {
 
   /// 炭水化物
-  /// int
-  int carbohydrateQuantity;
+  /// double
+  double carbohydrateQuantity;
 
   /// 食物繊維
-  /// int
-  int dietaryFiber;
+  /// double
+  double dietaryFiber;
 
   final formKey = GlobalKey<FormState>();
   final SugarsBloc sugarsBloc = SugarsBloc();
 
-  void calculate(int carbohydrateQuantity, int dietaryFiber) {
+  void calculate(double carbohydrateQuantity, double dietaryFiber) {
     sugarsBloc.calculate
         .add(CalculateFirstPatternEvent(carbohydrateQuantity, dietaryFiber));
   }
@@ -96,7 +96,7 @@ class CalculateSecondPatternPage extends StatelessWidget {
           hintText: '123',
           icon: Icon(Icons.device_unknown)),
       onSaved: (value) {
-        this.carbohydrateQuantity = int.parse(value);
+        this.carbohydrateQuantity = double.parse(value);
       },
     );
   }
@@ -117,7 +117,7 @@ class CalculateSecondPatternPage extends StatelessWidget {
           hintText: '123',
           icon: Icon(Icons.device_unknown)),
       onSaved: (value) {
-        this.dietaryFiber = int.parse(value);
+        this.dietaryFiber = double.parse(value);
       },
     );
   }
